pyproject.toml
[project]
name = "project-v-0-0-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "google-api-python-client>=2.171.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "markdown>=3.8",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
